Left ToF and Overcurrent Added
Need to work with Soohyun on the overcurrent, right now I'm reading an analog voltage from A12 assuming we'd put a resistor on the other side of the pin. Outside of that, all of the ToF I2C code works with the rest of our code. Implementing a second ToF sensor would be easy as well and I can add that blank code if we'd like.